requirements.txt
fastapi
openai
uvicorn
python-dotenv
mangum
